+++ src/geometry/CCPACSTransformation.h
@@ -41,6 +41,8 @@ public:
     TIGL_EXPORT CCPACSTransformation(CCPACSFuselageSection* parent, CTiglUIDManager* uidMgr);
     TIGL_EXPORT CCPACSTransformation(CCPACSExternalObject* parent, CTiglUIDManager* uidMgr);
     TIGL_EXPORT CCPACSTransformation(CCPACSGenericSystem* parent, CTiglUIDManager* uidMgr);
+    TIGL_EXPORT CCPACSTransformation(CCPACSVessel* parent, CTiglUIDManager* uidMgr);
+    TIGL_EXPORT CCPACSTransformation(CCPACSFuelTank* parent, CTiglUIDManager* uidMgr);
     TIGL_EXPORT CCPACSTransformation(CCPACSGenericGeometryComponent* parent, CTiglUIDManager* uidMgr);
     TIGL_EXPORT CCPACSTransformation(CCPACSLandingGearBase* parent, CTiglUIDManager* uidMgr);
     TIGL_EXPORT CCPACSTransformation(CCPACSNacelleSection* parent, CTiglUIDManager* uidMgr);
@@ -59,13 +61,17 @@ public:
     TIGL_EXPORT void setTranslation(const CTiglPoint& translation);
     TIGL_EXPORT void setTranslation(const CTiglPoint& translation, ECPACSTranslationType);
     TIGL_EXPORT void setRotation(const CTiglPoint& rotation);
+    TIGL_EXPORT void setRotationType(ECPACSTranslationType rotationType);
     TIGL_EXPORT void setScaling(const CTiglPoint& scale);
+    TIGL_EXPORT void setScalingType(ECPACSTranslationType scalingType);
     TIGL_EXPORT void setTransformationMatrix(const CTiglTransformation& matrix);
     
     TIGL_EXPORT CTiglPoint getTranslationVector() const;
     TIGL_EXPORT CTiglPoint getRotation() const;
     TIGL_EXPORT CTiglPoint getScaling() const;
     TIGL_EXPORT ECPACSTranslationType getTranslationType() const;
+    TIGL_EXPORT ECPACSTranslationType getScalingType() const;
+    TIGL_EXPORT ECPACSTranslationType getRotationType() const;
     TIGL_EXPORT CTiglTransformation getTransformationMatrix() const;
 
     // Helper function to set or reset a CCPACSTransformation.
@@ -87,6 +93,10 @@ private:
 
     // caches the transformation created from scaling, rotation and translation
     Cache<CTiglTransformation, CCPACSTransformation> _transformationMatrix;
+
+    // ABS_LOCAL per default only applies for translation, while scaling and rotation is global
+    ECPACSTranslationType _scalingType = ABS_GLOBAL;
+    ECPACSTranslationType _rotationType = ABS_GLOBAL;
 };
 
 } // namespace tigl